similarity/requirements.txt
biopython
rdkit
rdkit-pypi
uvicorn==0.30.6
fastapi==0.112.0
fastapi-cli==0.0.5
flatbuffers==24.3.25
slowapi==0.1.9